pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "news-trend-analyzer"
version = "0.1.0"
description = "A project to analyze news trends"
dependencies = [
    "python-dateutil",
    "feedparser",
    "psycopg",
    "psycopg[binary]",
    "textblob",
    "streamlit",
    "pandas"
]

[tool.setuptools.packages.find]
where = ["src"]